+++ smartimmo/src/main/java/com/gangdestrois/smartimmo/domain/portfolio/propertiesToFollow/port/PropertyToFollowApi.java
@@ -1,10 +1,10 @@
 package com.gangdestrois.smartimmo.domain.portfolio.propertiesToFollow.port;
 
 import com.gangdestrois.smartimmo.domain.portfolio.propertiesToFollow.model.PropertyToFollow;
-import com.gangdestrois.smartimmo.domain.property.entite.Property;
+import com.gangdestrois.smartimmo.domain.property.model.Property;
 import java.util.List;
 
 public interface PropertyToFollowApi {
-    List<PropertyToFollow> findAllByBuyerId(int buyerId);
-    List<Property> savePropertiesToFollowForBuyer(int buyerId);
+    List<PropertyToFollow> findAllByBuyerId(Long buyerId);
+    List<Property> savePropertiesToFollowForBuyer(Long buyerId);
 }
