+++ smartimmo/src/main/java/com/gangdestrois/smartimmo/infrastructure/jpa/repository/PropertyToFollowRepository.java
@@ -7,7 +7,6 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-
 import java.util.List;
 
 
@@ -16,7 +15,7 @@ public interface PropertyToFollowRepository extends JpaRepository<PropertyToFoll
     @Modifying
     @Transactional
     @Query(value = "delete from property_to_follow where fk_buyer = :buyerId", nativeQuery = true)
-    void deleteAllByBuyer_Id(@Param("buyerId") int buyerId);
+    void deleteAllByBuyer_Id(@Param("buyerId") Long buyerId);
 
-    List<PropertyToFollowEntity> findPropertyToFollowEntitiesByBuyer_Id(int buyerId);
+    List<PropertyToFollowEntity> findPropertyToFollowEntitiesByBuyer_Id(Long buyerId);
 }
